[MRG] Refact google_drive_downloader function (#473)
* refact google_drive_downloader function

* remove the original google_drive_downloader.py file
